It has been over a year since I last looked at this.
Right now I am going to start work on the parser for the SpiralV5 language (the Cuda backend) as the syntax is so obnoxious so of course I would look at the parser for GVGAI first.

It was the only large scala parser project that I did and it contains all my experiences of working in such a domain.

I won't make the language indentation sensitive at first, but it always an option once I am done with the first version. I actually rather prefer indentation sensitive languages like F# and Python over semicolonic ones.

As soon as I opened up this project a familiar sight of the Y Combinator greeted me.

I used it a year ago to make the parser call itself. I tried making it a recursive function a year ago and failed, but it seems I did learn a thing or two in the meantime. Back then I barely understood the Y Combinator, but now I can make in my sleep. I've found a few uses for it since then, but unlike here where I could have done without it, it was actually useful elsewhere.

I came really close to the solution here last time, but I got tripped up by partial function application.

In F# `let f x = g x` is not exactly the same as `let f = g` like it is in Haskell. The first expression is a function declaration, but the other is a statement that will get evaluated on the spot.

In `let rec sprites_up = many1Indents sprite sprites_up |>> SubSprites` if the F# compiler did not give me a type error, the expression would have gone into an infinite loop.

`let rec sprites_up x = (many1Indents sprite sprites_up |>> SubSprites) x` is the correct way of writing it.

Now let me do that pass over this code and recollect my memories so I can make the draft for the first version of the parser.